+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+- Add helper functions `isObject`, `isArray`, `isString`, `isNumber`, `isBoolean`
+
+---
+
 ## v0.5.0 – 2025-05-19
 
 - Introduce fallback method and FallbackSchema